Move sandbox to platform
