+++ src/vs/platform/sandbox/common/terminalSandbox.ts
